+++ xcm/xcm-builder/src/lib.rs
@@ -16,21 +16,40 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+#[cfg(test)]
+mod mock;
+#[cfg(test)]
+mod tests;
+
 mod location_conversion;
 pub use location_conversion::{
-	Account32Hash, ParentIsDefault, ChildParachainConvertsVia, SiblingParachainConvertsVia, AccountId32Aliases, AccountKey20Aliases,
+	Account32Hash, ParentIsDefault, ChildParachainConvertsVia, SiblingParachainConvertsVia, AccountId32Aliases,
+	AccountKey20Aliases, LocationInverter,
 };
 
 mod origin_conversion;
 pub use origin_conversion::{
 	SovereignSignedViaLocation, ParentAsSuperuser, ChildSystemParachainAsSuperuser, SiblingSystemParachainAsSuperuser,
-	ChildParachainAsNative, SiblingParachainAsNative, RelayChainAsNative, SignedAccountId32AsNative, SignedAccountKey20AsNative,
+	ChildParachainAsNative, SiblingParachainAsNative, RelayChainAsNative, SignedAccountId32AsNative,
+	SignedAccountKey20AsNative,
+};
+
+mod barriers;
+pub use barriers::{
+	TakeWeightCredit, AllowUnpaidExecutionFrom, AllowTopLevelPaidExecutionFrom, AllowKnownQueryResponses,
 };
 
 mod currency_adapter;
-mod fungibles_adapter;
 pub use currency_adapter::CurrencyAdapter;
+
+mod fungibles_adapter;
 pub use fungibles_adapter::FungiblesAdapter;
 
-#[deprecated="use `xcm-executor::traits::LocationInverter` instead"]
-pub use xcm_executor::traits::LocationInverter;
+mod weight;
+pub use weight::{FixedRateOfConcreteFungible, FixedWeightBounds};
+
+mod matches_fungible;
+pub use matches_fungible::{IsAbstract, IsConcrete};
+
+mod filter_asset_location;
+pub use filter_asset_location::{Case, NativeAsset};
